split data sources into separate files
