+++ src/module.js
@@ -13,7 +13,10 @@ import {SceneWorker} from '../src/scene_worker';
 import version from './utils/version';
 import log from 'loglevel';
 import Geo from './geo';
-import DataSource from './data_source';
+import DataSource from './sources/data_source';
+import './sources/geojson';
+import './sources/topojson';
+import './sources/mvt';
 import TileManager from './tile_manager';
 import GLSL from './gl/glsl';
 import ShaderProgram from './gl/shader_program';
